I checked this and this does get a speedup.

In the example I checked, the first epoch took 33.2 sec and subsequent epochs took about 24.3 sec.

After this PR, the first epoch takes 32.6 sec and subsequent epochs take about 23.7 sec